refactor: ChatRoomRepositoryImpl 엔티티 코틀린으로 변환 (#11)
- GroupChatResponse 에서 participantCount Int -> Long 타입 변경
- 코틀린으로 변환 하니까 toInt() 메서드를 사용할 수 없음.

+++ backend/src/main/java/com/app/backend/domain/group/dto/response/GroupChatResponse.java
@@ -1,5 +1,5 @@
 package com.app.backend.domain.group.dto.response;
 
-public record GroupChatResponse(Long groupId, String groupName, int participantCount) {
+public record GroupChatResponse(Long groupId, String groupName, Long participantCount) {
 
 }
